Move metadata.txt and __init__.py into arches_project dir #58
