--- __init__.py
@@ -1,36 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-/***************************************************************************
- ArchesProject
-                                 A QGIS plugin
- This plugin links QGIS to an Arches project.
- Generated by Plugin Builder: http://g-sherman.github.io/Qgis-Plugin-Builder/
-                             -------------------
-        begin                : 2023-09-15
-        copyright            : (C) 2023 by Knowledge Integration
-        email                : samuel.scandrett@k-int.co.uk
-        git sha              : $Format:%H$
- ***************************************************************************/
-
-/***************************************************************************
- *                                                                         *
- *   This program is free software; you can redistribute it and/or modify  *
- *   it under the terms of the GNU General Public License as published by  *
- *   the Free Software Foundation; either version 2 of the License, or     *
- *   (at your option) any later version.                                   *
- *                                                                         *
- ***************************************************************************/
- This script initializes the plugin, making it known to QGIS.
-"""
-
-
-# noinspection PyPep8Naming
-def classFactory(iface):  # pylint: disable=invalid-name
-    """Load ArchesProject class from file ArchesProject.
-
-    :param iface: A QGIS interface instance.
-    :type iface: QgsInterface
-    """
-    #
-    from .arches_project.arches_project import ArchesProject
-    return ArchesProject(iface)
